requirements.txt
ccxt>=4.0.0
pandas>=1.5.0
requests>=2.28.0
python-dotenv>=0.19.0
flask>=2.0.0
flask-socketio>=5.0.0
python-socketio>=5.0.0
eventlet>=0.33.0
google-genai>=0.1.0
